refactor(desktop-tauri): split tauri lib into modules
